# portal-api-develop/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "iemanager",
    "description": "IEManger Application Backend APIs",
    "contact": {}
  },
  "host": "",
  "basePath": "/iemanager",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/iems/{iemsId}/devices": {
      "get": {
        "description": "List all devices that onboarded on specific iems",
        "summary": "Get iems devices",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiV1Users100Iems10DevicesGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "iemsId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/api/v1/iems/{iemsId}/devices/checkout": {
      "post": {
        "description": "Checkout License for new device onboarding to specific iems",
        "summary": "Check out device license",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiV1Users100Iems10DevicesCheckoutPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          }, 
          {
            "name": "iemsId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckoutdevicelicenseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/v1/iems/{iemsId}/devices/check-in": {
      "post": {
        "description": "Check in license per specific iems ",
        "summary": "Check in license device",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiV1Users100Iems10DevicesCheckInPost",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          },
          
          {
            "name": "iemsId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckinlicensedeviceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/v1/iems/{iemsId}/devices/check-in-devices": {
      "post": {
        "description": "Check in all licence that consumed by specific IEMS",
        "summary": "Check-in all devices per IEMS",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiV1Users100Iems100DevicesCheckInDevicesGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          },
          
          {
            "name": "iemsId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/api/v1/transactions/devices/{fromDate}": {
      "get": {
        "description": "List all device transactions",
        "summary": "Get user IE devices transactions history",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiUsersUserIdDeviceTransactionsGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fromDate",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "toDate",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CheckoutdevicelicenseRequest": {
      "title": "CheckoutdevicelicenseRequest",
      "example": {
        "serialNo": "Abc091231",
        "name" : "device_name"
      },
      "type": "object",
      "properties": {
        "device_sn": {
          "type": "string"
        },
        "device_name": {
          "type": "string"
        },
        "iems_hostname": {
          "type": "string"
        }
      },
      "required": [
        "serialNo",
        "name"
      ]
    },
    "CheckinlicensedeviceRequest": {
      "title": "CheckinlicensedeviceRequest",
      "example": {
        "serialNo": "Device-PRE-001"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "serialNo": {
          "type": "string"
        },
        "device_name": {
          "type": "string"
        },
        "license_status": {
          "type": "string"
        },
        "subscriptionDate": {
          "type": "integer",
          "format": "int64"
        }
      },
      "required": [
        "device_name"
      ]
    },
...
  },
...
}